+++ src/lib/patchkit_error.rb
@@ -1,0 +1,4 @@
+module PatchKitTools
+  class CommandLineError < StandardError
+  end
+end