+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- Patchkit tools online documentation url in main help message
 - Add command help command line to main help message
+- Support for internal tools by PK_TOOLS_INTERNAL environment variable.
+
+## [0.9.2]
+### Added
+- Patchkit tools online documentation url in main help message
 
 ### Changed
 - Config file is now optional